pyproject.toml
[tool.poetry]
name = "django-boilerplate"
version = "0.1.0"
description = ""
authors = ["igorsaulo <igorsaulo01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
